# BackendApplication/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: WrocłArts API
  description: REST API documentation for the WrocłArts system managing artistic activities in Wrocław.
  termsOfService: http://www.example.com/terms/
  contact:
    email: contact@maxime-luce.fr
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/WroclArts
    description: Local Spring Boot Development Server
...
paths:
  /auth/login:
    post:
      tags:
        - user
      summary: Authenticate user and get token
      operationId: loginUser
      requestBody:
        description: User credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginRequest'
            application/xml:
              schema:
                $ref: '#/components/schemas/LoginRequest'
        400:
          description: Invalid credentials
          content: {}
      x-codegen-request-body-name: loginRequest
  /users/register:
    post:
      tags:
        - user
      summary: Register a new visitor account
      operationId: registerUser
      requestBody:
        description: User data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        200:
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistration'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserRegistration'
        400:
          description: Invalid input
          content: {}
      x-codegen-request-body-name: userRegistration
...
  /reviews:
    post:
      tags:
        - user
      summary: Leave a review for an event or organization
      operationId: postReview
      requestBody:
        description: Review details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
          application/xml:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        200:
          description: Review posted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
            application/xml:
              schema:
                $ref: '#/components/schemas/Review'
      security:
        - wroclarts_auth:
            - write:events
      x-codegen-request-body-name: review
  /support/tickets:
    post:
      tags:
        - user
      summary: Submit a support ticket to platform administrators
      operationId: submitSupportTicket
      requestBody:
        description: Support ticket details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportTicket'
          application/xml:
            schema:
              $ref: '#/components/schemas/SupportTicket'
        required: true
      responses:
        200:
          description: Support ticket submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicket'
            application/xml:
              schema:
                $ref: '#/components/schemas/SupportTicket'
      x-codegen-request-body-name: supportTicket
...
components:
  schemas:
    LoginRequest:
      required:
        - email
        - password
      type: object
      xml:
        name: LoginRequest
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserRegistration:
      required:
        - email
        - password
      type: object
      xml:
        name: UserRegistration
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        username:
          type: string
    Review:
      required:
        - rating
        - comment
      type: object
      xml:
        name: Review
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    SupportTicket:
      required:
        - subject
        - message
      type: object
      xml:
        name: SupportTicket
      properties:
        subject:
          type: string
        message:
          type: string
...
  securitySchemes:
    wroclarts_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost/oauth/dialog
          scopes:
            write:events: modify events records in your account
            read:events: read your events records
            write:organizations: modify organization records
            read:organizations: read organization records
            write:products: modify shop products
            read:products: read shop products
            write:orders: modify order status
            read:orders: read order status
            admin:platform: exclusive platform administrator actions (moderation, verification)